+++ src/db_building/__init__.py
@@ -1,10 +1,4 @@
-from .build_database import (
-    build_vector_database,
-    reload_database_from_local_dir,
-)
-from .loading import (
-    load_retriever
-)
+from .build_database import build_vector_database, load_retriever, reload_database_from_local_dir
 
 __all__ = [
     "build_vector_database",
